+++ packages/2d/src/Components/Geometry.ts
@@ -2,6 +2,7 @@ import { useType } from "@hex-engine/core";
 import { Vector, Shape } from "../Models";
 import {
   useInspectorHoverOutline,
+  useInspectorSelectEntity,
   useEntityTransforms,
   useDebugOverlayDrawTime,
 } from "../Hooks";
@@ -40,6 +41,7 @@ function Geometry<S extends Shape>({
 
   useDebugOverlayDrawTime();
   useInspectorHoverOutline(() => geometry.shape);
+  useInspectorSelectEntity(() => geometry);
 
   return geometry;
 }