+++ packages/inspector/src/App.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { Entity, RunLoop } from "@hex-engine/core";
-import { StateKey, useStateTree } from "react-state-tree";
 import Tree from "./Tree";
 import Controls from "./Controls";
 import PausedOverlay from "./PausedOverlay";
@@ -11,16 +10,28 @@ export default function App({
   entity,
   runLoop,
   error,
+  onExpand,
+  getExpanded,
+  getSelectedEntity,
   isHovered,
+  isOpen,
+  toggleOpen,
+  isSelectMode,
+  toggleSelectMode,
 }: {
   entity: Entity;
+  getSelectedEntity: () => null | Entity;
   runLoop: RunLoopAPI | null;
   error: Error | null;
+  onExpand: (path: Array<string | number>, expand: boolean) => void;
+  getExpanded: (path: Array<string | number>) => boolean;
   isHovered: boolean;
+  isOpen: boolean;
+  toggleOpen: () => void;
+  isSelectMode: boolean;
+  toggleSelectMode: () => void;
 }) {
   let ent = entity;
-
-  const [open, setOpen] = useStateTree(false, "open");
 
   return (
     <div
@@ -45,21 +56,21 @@ export default function App({
             : null),
         }}
       >
-        <StateKey value="controls">
-          {runLoop && !open ? (
-            <Controls
-              isOpen={open}
-              toggleOpen={() => setOpen(!open)}
-              runLoop={runLoop}
-              error={error}
-            />
-          ) : null}
-        </StateKey>
+        {runLoop && !isOpen ? (
+          <Controls
+            isSelectMode={isSelectMode}
+            toggleSelectMode={toggleSelectMode}
+            isOpen={isOpen}
+            toggleOpen={toggleOpen}
+            runLoop={runLoop}
+            error={error}
+          />
+        ) : null}
       </div>
       {runLoop && runLoop.isPaused() && runLoop.frameNumber === 0 ? (
         <PausedOverlay runLoop={runLoop} />
       ) : null}
-      {open ? (
+      {isOpen ? (
         <div
           style={{
             position: "fixed",
@@ -84,21 +95,27 @@ export default function App({
               : null),
           }}
         >
-          <StateKey value="controls">
-            {runLoop ? (
-              <Controls
-                isOpen={open}
-                toggleOpen={() => setOpen(!open)}
-                runLoop={runLoop}
-                error={error}
-              />
-            ) : null}
-          </StateKey>
-          <StateKey value="tree">
-            <div style={{ flexBasis: "100%" }}>
-              <Tree name="root" data={ent} parent={null} />
-            </div>
-          </StateKey>
+          {runLoop ? (
+            <Controls
+              isSelectMode={isSelectMode}
+              toggleSelectMode={toggleSelectMode}
+              isOpen={isOpen}
+              toggleOpen={toggleOpen}
+              runLoop={runLoop}
+              error={error}
+            />
+          ) : null}
+          <div style={{ flexBasis: "100%" }}>
+            <Tree
+              name="root"
+              data={ent}
+              parent={null}
+              path={["root"]}
+              onExpand={onExpand}
+              getExpanded={getExpanded}
+              getSelectedEntity={getSelectedEntity}
+            />
+          </div>
         </div>
       ) : null}
     </div>