+++ packages/2d/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+# 0.3.2
+
+## Upgraded `@hex-engine/2d` and `@hex-engine/inspector` to 0.3.2 (non-breaking change)
+
+This adds a feature to select rendered entities for inspection in the Inspector tree if the Inspector is in entity select mode.
+
 # 0.3.1
 
 ## Upgraded `@hex-engine/core` to 0.3.1 (non-breaking change)